[Server] Use program startup approach. Removed not needed code
